backend/requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_jwt_extended
psycopg2-binary
google-cloud-secret-manager
chromadb
requests
gunicorn